shallow reorganization of the crate, completely remove legacy server stuff except for the compatibility adapter
