+++ Cargo.toml
@@ -40,5 +40,3 @@ parking_lot = "0.12.0"
 nanorpc = "0.1.3"
 serde_json = "1.0.83"
 
-[features]
-server=[]
